+++ src/Managers/AuthManager/index.js
